Create metabolism
idk, there are a lot of places where the code could be a lot simpler (e.g. n_O2_out = 0 instead of doing the accumulation equation, since we know all oxygen is consumed) but I wasn't really sure if we're allowed to do shortcuts like that? so I just did the calculations and it should turn out to be 0 anyways :/